requirements.txt
mysql.connector
mysql-connector-python
pyodbc
flask
flask_cors
dotenv
openai